Add initial provider structure
Add a main.go that has the terraform boiler plate all plugins need.
Add provider.go which will initialise the provider to connect with elasticsearch.
The provider accepts a variable `url` which is the elasticsearch URL.
The provider has a configure function that sets up an elasticsearch client, this is available in the resource methods.